Add new flag to set the default content type to something else than the generic application/octect-stream.
The module "send" is the module responsible to set the content type header in the response, it uses the mime module 1.2.6 to determine the appropriate mime type. The module "mime" exposes mime.default_type by using the same version in both "serve" and "send" I am able to change the default mime type return in "send".